商品削除機能２

+++ app/controllers/items_controller.rb
@@ -36,9 +36,11 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+  if current_user.id == @item.user_id
     if @item.destroy
       redirect_to root_path
     end
+  end
   end
   private
